add fetching cached data, add hourly and daily forecast

+++ WeatherApp/Persistance/CurrentForecastEntity+CoreDataProperties.swift
@@ -26,7 +26,7 @@ extension CurrentForecastEntity {
     @NSManaged public var condition: String?
     @NSManaged public var feelC: Double
     @NSManaged public var fellF: Double
-    @NSManaged public var conditionCode: Int16
+    @NSManaged public var conditionCode: Int32
     @NSManaged public var forecastEntity: ForecastEntity?
 
 }
